+++ modules/gapi/CMakeLists.txt
@@ -186,8 +186,6 @@ if(HAVE_GAPI_ONEVPL)
         src/streaming/onevpl/engine/processing_engine_base.cpp
         src/streaming/onevpl/engine/decode/decode_engine_legacy.cpp
         src/streaming/onevpl/engine/decode/decode_session.cpp
-        src/streaming/onevpl/engine/decode/decode_engine_async.cpp
-        src/streaming/onevpl/engine/decode/decode_session_async.cpp
         src/streaming/onevpl/onevpl_utils.cpp
         src/streaming/onevpl/onevpl_cfg_params_parser.cpp
     )
