+++ modules/gapi/src/streaming/onevpl/engine/decode/decode_engine_legacy.cpp
@@ -20,7 +20,7 @@ namespace cv {
 namespace gapi {
 namespace wip {
 /* UTILS */
-mfxU32 GetSurfaceSize(mfxU32 FourCC, mfxU32 width, mfxU32 height) {
+mfxU32 GetSurfaceSize_(mfxU32 FourCC, mfxU32 width, mfxU32 height) {
     mfxU32 nbytes = 0;
 
     mfxU32 half_width = width / 2;
@@ -45,7 +45,7 @@ mfxU32 GetSurfaceSize(mfxU32 FourCC, mfxU32 width, mfxU32 height) {
     return nbytes;
 }
 
-surface_ptr_t create_surface_RGB4(mfxFrameInfo frameInfo,
+surface_ptr_t create_surface_RGB4_(mfxFrameInfo frameInfo,
                                   std::shared_ptr<void> out_buf_ptr,
                                   size_t out_buf_ptr_offset,
                                   size_t out_buf_size)
@@ -78,7 +78,7 @@ surface_ptr_t create_surface_RGB4(mfxFrameInfo frameInfo,
     return Surface::create_surface(std::move(handle), out_buf_ptr);
 }
 
-surface_ptr_t create_surface_other(mfxFrameInfo frameInfo,
+surface_ptr_t create_surface_other_(mfxFrameInfo frameInfo,
                                    std::shared_ptr<void> out_buf_ptr,
                                    size_t out_buf_ptr_offset,
                                    size_t out_buf_size)
@@ -110,69 +110,95 @@ surface_ptr_t create_surface_other(mfxFrameInfo frameInfo,
     return Surface::create_surface(std::move(handle), out_buf_ptr);
 }
 
-VPLLegacyDecodeEngine::VPLLegacyDecodeEngine(std::unique_ptr<VPLAccelerationPolicy>&& accel)
+VPLLegacyDecodeEngineAsync::VPLLegacyDecodeEngineAsync(std::unique_ptr<VPLAccelerationPolicy>&& accel)
  : ProcessingEngineBase(std::move(accel)) {
 
     GAPI_LOG_INFO(nullptr, "Create Legacy Decode Engine");
     create_pipeline(
-        // 1) Reade File
+        // 1) Read File
         [this] (EngineSession& sess) -> ExecutionStatus
         {
-            LegacyDecodeSession &my_sess = static_cast<LegacyDecodeSession&>(sess);
+            LegacyDecodeSessionAsync &my_sess = static_cast<LegacyDecodeSessionAsync&>(sess);
             my_sess.last_status = ReadEncodedStream(my_sess.stream, my_sess.data_provider);
             if (my_sess.last_status != MFX_ERR_NONE) {
                 my_sess.data_provider.reset(); //close source
             }
             return ExecutionStatus::Continue;
         },
-        // 2) enqueue ASYNC decode
+        // 2) enqueue ASYNC decode operation
         [this] (EngineSession& sess) -> ExecutionStatus
         {
-            LegacyDecodeSession &my_sess = static_cast<LegacyDecodeSession&>(sess);
+            LegacyDecodeSessionAsync &my_sess = static_cast<LegacyDecodeSessionAsync&>(sess);
 
+            // prepare sync object for new surface
+            LegacyDecodeSessionAsync::op_handle_t sync_pair{};
+
+            // queue qecode operation
             my_sess.last_status =
                     MFXVideoDECODE_DecodeFrameAsync(my_sess.session,
                                                     my_sess.last_status == MFX_ERR_NONE
                                                         ? &my_sess.stream
                                                         : nullptr, /* No more data to read, start decode draining mode*/
                                                     my_sess.procesing_surface_ptr.lock()->get_handle(),
-                                                    &my_sess.output_surface_ptr,
-                                                    &my_sess.sync);
+                                                    &sync_pair.second,
+                                                    &sync_pair.first);
 
-            if (!pending_frames.empty())
-            {
-                auto data = pending_frames.front();
-                pending_frames.pop();
-                ready_frames.push(std::move(data));
+            // process wait-like statuses in-place
+            while (my_sess.last_status == MFX_ERR_MORE_SURFACE ||
+                   my_sess.last_status == MFX_WRN_DEVICE_BUSY) {
+                try {
+                    if (my_sess.last_status == MFX_ERR_MORE_SURFACE) {
+                        my_sess.swap_surface(*this);
+                    }
+                    my_sess.last_status =
+                    MFXVideoDECODE_DecodeFrameAsync(my_sess.session,
+                                                   &my_sess.stream,
+                                                    my_sess.procesing_surface_ptr.lock()->get_handle(),
+                                                    &sync_pair.second,
+                                                    &sync_pair.first);
+
+                } catch (const std::exception& ex) {
+                    GAPI_LOG_WARNING(nullptr, "[" << my_sess.session << "] error: " << ex.what() <<
+                                            "Abort");
+                }
+            }
+
+            if (my_sess.last_status == MFX_ERR_NONE) {
+                my_sess.sync_queue.emplace(sync_pair);
+            } else if (MFX_ERR_MORE_DATA != my_sess.last_status) /* suppress MFX_ERR_MORE_DATA warning */ {
+                GAPI_LOG_WARNING(nullptr, "pending ops count: " << my_sess.sync_queue.size() <<
+                                        ", sync id: " << sync_pair.first <<
+                                        ", status: " << mfxstatus_to_string(my_sess.last_status));
             }
             return ExecutionStatus::Continue;
         },
         // 3) Wait for ASYNC decode result
         [this] (EngineSession& sess) -> ExecutionStatus
         {
-            if (sess.last_status == MFX_ERR_NONE) // Got 1 decoded frame
+            LegacyDecodeSessionAsync& my_sess = static_cast<LegacyDecodeSessionAsync&>(sess);
+            if (!my_sess.sync_queue.empty()) // FIFO: check the oldest async operation complete
             {
-                do {
-                    //TODO try to extract TIMESTAMP
-                    sess.last_status = MFXVideoCORE_SyncOperation(sess.session, sess.sync, 100);
-                    if (MFX_ERR_NONE == sess.last_status) {
+                LegacyDecodeSessionAsync::op_handle_t& pending_op = my_sess.sync_queue.front();
+                sess.last_status = MFXVideoCORE_SyncOperation(sess.session, pending_op.first, 0);
 
-                        LegacyDecodeSession& my_sess = static_cast<LegacyDecodeSession&>(sess);
-                        on_frame_ready(my_sess);
-                    }
-                } while (sess.last_status == MFX_WRN_IN_EXECUTION);
+                GAPI_LOG_DEBUG(nullptr, "pending ops count: " << my_sess.sync_queue.size() <<
+                                        ", sync id:  " << pending_op.first <<
+                                        ", status: " << mfxstatus_to_string(my_sess.last_status));
+                if (MFX_ERR_NONE == sess.last_status) {
+                    on_frame_ready(my_sess, pending_op.second);
+                }
             }
             return ExecutionStatus::Continue;
         },
         // 4) Falls back on generic status procesing
         [this] (EngineSession& sess) -> ExecutionStatus
         {
-            return this->process_error(sess.last_status, static_cast<LegacyDecodeSession&>(sess));
+            return this->process_error(sess.last_status, static_cast<LegacyDecodeSessionAsync&>(sess));
         }
     );
 }
 
-void VPLLegacyDecodeEngine::initialize_session(mfxSession mfx_session,
+void VPLLegacyDecodeEngineAsync::initialize_session(mfxSession mfx_session,
                                          DecoderParams&& decoder_param,
                                          std::shared_ptr<IDataProvider> provider)
 {
@@ -185,7 +211,7 @@ void VPLLegacyDecodeEngine::initialize_session(mfxSession mfx_session,
                             ", mfxFrameAllocRequest.NumFrameSuggested: " << decRequest.NumFrameSuggested <<
                             ", mfxFrameAllocRequest.Type: " << decRequest.Type);
 
-    mfxU32 singleSurfaceSize = GetSurfaceSize(decoder_param.param.mfx.FrameInfo.FourCC,
+    mfxU32 singleSurfaceSize = GetSurfaceSize_(decoder_param.param.mfx.FrameInfo.FourCC,
                                               decoder_param.param.mfx.FrameInfo.Width,
                                               decoder_param.param.mfx.FrameInfo.Height);
     if (!singleSurfaceSize) {
@@ -200,9 +226,9 @@ void VPLLegacyDecodeEngine::initialize_session(mfxSession mfx_session,
             [&frameInfo] (std::shared_ptr<void> out_buf_ptr, size_t out_buf_ptr_offset,
                           size_t out_buf_size) -> surface_ptr_t {
                 return (frameInfo.FourCC == MFX_FOURCC_RGB4) ?
-                        create_surface_RGB4(frameInfo, out_buf_ptr, out_buf_ptr_offset,
+                        create_surface_RGB4_(frameInfo, out_buf_ptr, out_buf_ptr_offset,
                                             out_buf_size) :
-                        create_surface_other(frameInfo, out_buf_ptr, out_buf_ptr_offset,
+                        create_surface_other_(frameInfo, out_buf_ptr, out_buf_ptr_offset,
                                              out_buf_size);};
 
     //TODO Configure preallocation size (how many frames we can hold)
@@ -213,8 +239,8 @@ void VPLLegacyDecodeEngine::initialize_session(mfxSession mfx_session,
                                                          surface_creator);
 
     // create session
-    std::shared_ptr<LegacyDecodeSession> sess_ptr =
-                register_session<LegacyDecodeSession>(mfx_session,
+    std::shared_ptr<LegacyDecodeSessionAsync> sess_ptr =
+                register_session<LegacyDecodeSessionAsync>(mfx_session,
                                                       std::move(decoder_param),
                                                       provider);
 
@@ -223,34 +249,39 @@ void VPLLegacyDecodeEngine::initialize_session(mfxSession mfx_session,
     sess_ptr->swap_surface(*this);
 }
 
-ProcessingEngineBase::ExecutionStatus VPLLegacyDecodeEngine::execute_op(operation_t& op, EngineSession& sess) {
+ProcessingEngineBase::ExecutionStatus VPLLegacyDecodeEngineAsync::execute_op(operation_t& op, EngineSession& sess) {
     return op(sess);
 }
 
-void VPLLegacyDecodeEngine::on_frame_ready(LegacyDecodeSession& sess)
+void VPLLegacyDecodeEngineAsync::on_frame_ready(LegacyDecodeSessionAsync& sess, mfxFrameSurface1* ready_surface)
 {
     GAPI_LOG_DEBUG(nullptr, "[" << sess.session << "], frame ready");
 
     // manage memory ownership rely on acceleration policy
     auto frame_adapter = acceleration_policy->create_frame_adapter(sess.decoder_pool_id,
-                                                                   sess.output_surface_ptr);
-    pending_frames.emplace(cv::MediaFrame(std::move(frame_adapter)), sess.generate_frame_meta());
+                                                                   ready_surface);
+    ready_frames.emplace(cv::MediaFrame(std::move(frame_adapter)), sess.generate_frame_meta());
+
+    // pop ready sync onject
+    sess.sync_queue.pop();
 }
 
-ProcessingEngineBase::ExecutionStatus VPLLegacyDecodeEngine::process_error(mfxStatus status, LegacyDecodeSession& sess)
+ProcessingEngineBase::ExecutionStatus VPLLegacyDecodeEngineAsync::process_error(mfxStatus status, LegacyDecodeSessionAsync& sess)
 {
     GAPI_LOG_DEBUG(nullptr, "status: " << mfxstatus_to_string(status));
 
     switch (status) {
         case MFX_ERR_NONE:
+        {
+            // prepare sync object for new surface
             try {
                 sess.swap_surface(*this);
+                return ExecutionStatus::Continue;
             } catch (const std::exception& ex) {
                 GAPI_LOG_WARNING(nullptr, "[" << sess.session << "] error: " << ex.what() <<
                                           "Abort");
-                ExecutionStatus::Failed;
             }
-            return ExecutionStatus::Continue;
+        }
         case MFX_ERR_MORE_DATA: // The function requires more bitstream at input before decoding can proceed
             if (!sess.data_provider || sess.data_provider->empty()) {
                 // No more data to drain from decoder, start encode draining mode
@@ -276,27 +307,27 @@ ProcessingEngineBase::ExecutionStatus VPLLegacyDecodeEngine::process_error(mfxSt
         case MFX_ERR_DEVICE_LOST:
             // For non-CPU implementations,
             // Cleanup if device is lost
-            GAPI_DbgAssert(false && "VPLLegacyDecodeEngine::process_error - "
+            GAPI_DbgAssert(false && "VPLLegacyDecodeEngineAsync::process_error - "
                                     "MFX_ERR_DEVICE_LOST is not processed");
             break;
         case MFX_WRN_DEVICE_BUSY:
             // For non-CPU implementations,
             // Wait a few milliseconds then try again
-            GAPI_DbgAssert(false && "VPLLegacyDecodeEngine::process_error - "
+            GAPI_DbgAssert(false && "VPLLegacyDecodeEngineAsync::process_error - "
                                     "MFX_WRN_DEVICE_BUSY is not processed");
             break;
         case MFX_WRN_VIDEO_PARAM_CHANGED:
             // The decoder detected a new sequence header in the bitstream.
             // Video parameters may have changed.
             // In external memory allocation case, might need to reallocate the output surface
-            GAPI_DbgAssert(false && "VPLLegacyDecodeEngine::process_error - "
+            GAPI_DbgAssert(false && "VPLLegacyDecodeEngineAsync::process_error - "
                                     "MFX_WRN_VIDEO_PARAM_CHANGED is not processed");
             break;
         case MFX_ERR_INCOMPATIBLE_VIDEO_PARAM:
             // The function detected that video parameters provided by the application
             // are incompatible with initialization parameters.
             // The application should close the component and then reinitialize it
-            GAPI_DbgAssert(false && "VPLLegacyDecodeEngine::process_error - "
+            GAPI_DbgAssert(false && "VPLLegacyDecodeEngineAsync::process_error - "
                                     "MFX_ERR_INCOMPATIBLE_VIDEO_PARAM is not processed");
             break;
         case MFX_ERR_REALLOC_SURFACE:
@@ -304,11 +335,20 @@ ProcessingEngineBase::ExecutionStatus VPLLegacyDecodeEngine::process_error(mfxSt
             // mfxInfoMFX::EnableReallocRequest was set to ON during initialization.
             // This applies to external memory allocations and should not be expected for
             // a simple internal allocation case like this
-            GAPI_DbgAssert(false && "VPLLegacyDecodeEngine::process_error - "
+            GAPI_DbgAssert(false && "VPLLegacyDecodeEngineAsync::process_error - "
                                     "MFX_ERR_REALLOC_SURFACE is not processed");
             break;
+        case MFX_WRN_IN_EXECUTION:
+            try {
+                sess.swap_surface(*this);
+                return ExecutionStatus::Continue;
+            } catch (const std::exception& ex) {
+                GAPI_LOG_WARNING(nullptr, "[" << sess.session << "] error: " << ex.what() <<
+                                          "Abort");
+            }
         default:
-            GAPI_LOG_WARNING(nullptr, "Unknown status code: " << mfxstatus_to_string(status));
+            GAPI_LOG_WARNING(nullptr, "Unknown status code: " << mfxstatus_to_string(status) <<
+                                      ", decoded frames: " << sess.decoded_frames_count);
             break;
     }
