+++ modules/gapi/src/streaming/onevpl/engine/decode/decode_engine_legacy.hpp
@@ -4,8 +4,8 @@
 //
 // Copyright (C) 2021 Intel Corporation
 
-#ifndef GAPI_STREAMING_ONVPL_ENGINE_DECODE_DECODE_ENGINE_LEGACY_HPP
-#define GAPI_STREAMING_ONVPL_ENGINE_DECODE_DECODE_ENGINE_LEGACY_HPP
+#ifndef GAPI_STREAMING_ONVPL_ENGINE_DECODE_DECODE_ENGINE_ASYNC_HPP
+#define GAPI_STREAMING_ONVPL_ENGINE_DECODE_DECODE_ENGINE_ASYNC_HPP
 #include <stdio.h>
 #include <memory>
 
@@ -22,28 +22,26 @@ namespace gapi {
 namespace wip {
 
 
-class LegacyDecodeSession;
+class LegacyDecodeSessionAsync;
 struct DecoderParams;
 struct IDataProvider;
 struct VPLAccelerationPolicy;
 
-class VPLLegacyDecodeEngine : public ProcessingEngineBase {
+class VPLLegacyDecodeEngineAsync : public ProcessingEngineBase {
 public:
 
-    VPLLegacyDecodeEngine(std::unique_ptr<VPLAccelerationPolicy>&& accel);
+    VPLLegacyDecodeEngineAsync(std::unique_ptr<VPLAccelerationPolicy>&& accel);
     void initialize_session(mfxSession mfx_session, DecoderParams&& decoder_param,
                             std::shared_ptr<IDataProvider> provider) override;
 
 private:
     ExecutionStatus execute_op(operation_t& op, EngineSession& sess) override;
-    ExecutionStatus process_error(mfxStatus status, LegacyDecodeSession& sess);
+    ExecutionStatus process_error(mfxStatus status, LegacyDecodeSessionAsync& sess);
 
-    void on_frame_ready(LegacyDecodeSession& sess);
-
-    frames_container_t pending_frames;
+    void on_frame_ready(LegacyDecodeSessionAsync& sess, mfxFrameSurface1* ready_surface);
 };
 } // namespace wip
 } // namespace gapi
 } // namespace cv
 #endif // HAVE_ONEVPL
-#endif // GAPI_STREAMING_ONVPL_ENGINE_DECODE_DECODE_ENGINE_LEGACY_HPP
+#endif // GAPI_STREAMING_ONVPL_ENGINE_DECODE_DECODE_ENGINE_ASYNC_HPP
